:bug: Fix Issue #1
:warning: assigning `lspConfig.ycm-lsp-server.name` to the primary file
type, AKA `:echo &filetype` as reported by Vim, means it is possible for
multiple servers to run for the same language and odd behavior to ensue.

Future Nix checks may need added to the root `./default.nix` file of
this repo to at least warn when such things are likely.